docs: addressed request changes
