+++ Superlista/Models/LocalModels/CategoryModel.swift
@@ -14,13 +14,3 @@ struct CategoryModel: Decodable, Hashable, Encodable {
     }
 }
 
-extension Array where Element: Equatable {
-    var unique: [Element] {
-        var uniqueValues: [Element] = []
-        forEach { item in
-            guard !uniqueValues.contains(item) else { return }
-            uniqueValues.append(item)
-        }
-        return uniqueValues
-    }
-}
